requirements.txt
PyGithub~=1.55
GitPython
nose
requests~=2.27.1
xmltodict~=0.12.0
psycopg2
pyOptional~=1.1.2
azure-cosmos
arrow~=1.2.2
gremlinpython==3.5.1
future
coloredlogs
backoff~=1.11.1
pyyaml~=6.0
azure-storage-blob
azure-storage-queue
azure-servicebus
pandas~=1.4.2
python-dotenv~=0.20.0
python-dateutil~=2.8.2
setuptools~=60.2.0